Fixed dialog issues, changed the way projectiles get damage values and added a check for using your melee

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void Hurt(float damage, GameObject damageSource);
+    public void Hurt(int damage, GameObject damageSource);
     public void Kill();
 }
